Prepare ReloadDetection to be shared
