+++ robozonky-common/src/main/java/com/github/robozonky/common/async/ReloadableImpl.java
@@ -17,14 +17,10 @@
 package com.github.robozonky.common.async;
 
 import java.time.Duration;
-import java.time.Instant;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicReference;
-import java.util.function.BooleanSupplier;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-import com.github.robozonky.internal.util.DateUtil;
 import io.vavr.control.Either;
 import io.vavr.control.Try;
 import org.slf4j.Logger;
@@ -87,64 +83,5 @@ final class ReloadableImpl<T> implements Reloadable<T> {
                     LOGGER.debug("Reloaded {}, new value is {}.", this, v);
                 })
                 .toEither();
-    }
-
-    private interface ReloadDetection extends BooleanSupplier {
-
-        void markReloaded();
-
-        void forceReload();
-    }
-
-    private static final class ManualReload implements ReloadDetection {
-
-        private final AtomicBoolean needsReload = new AtomicBoolean(true);
-
-        @Override
-        public boolean getAsBoolean() {
-            return needsReload.get();
-        }
-
-        @Override
-        public void markReloaded() {
-            needsReload.set(false);
-            LOGGER.trace("Marked reloaded on {}.", this);
-        }
-
-        @Override
-        public void forceReload() {
-            needsReload.set(true);
-            LOGGER.trace("Forcing reload on {}.", this);
-        }
-    }
-
-    private static final class TimeBasedReload implements ReloadDetection {
-
-        private final AtomicReference<Instant> lastReloaded;
-        private final Duration reloadAfter;
-
-        public TimeBasedReload(final Duration reloadAfter) {
-            this.reloadAfter = reloadAfter;
-            lastReloaded = new AtomicReference<>();
-        }
-
-        @Override
-        public boolean getAsBoolean() {
-            final Instant lastReloadedInstant = lastReloaded.get();
-            return lastReloadedInstant == null ||
-                    lastReloadedInstant.plus(reloadAfter).isBefore(DateUtil.now());
-        }
-
-        @Override
-        public void markReloaded() {
-            lastReloaded.set(DateUtil.now());
-            LOGGER.trace("Marked reloaded on {}.", this);
-        }
-
-        @Override
-        public void forceReload() {
-            lastReloaded.set(null);
-            LOGGER.trace("Forcing reload on {}.", this);
-        }
     }
 }
